Refactor time management
